Create CMP(Q2)Decimation.f
Consider an infinitely long periodic chain of atoms. Each atomic site is assigned an on-site potential equal to zero. The nearest neighbor hopping integral can be chosen to be `1’. Using the difference equations and the decimation method, obtain the Green’s function for any site and hence calculate and plot the Local Density of States. Take the imaginary part as 0.01 and add it to the energy E. Compare your result with part 1 above.